pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "cachetools>=6.2.0",
    "celery>=5.5.3",
    "fastapi>=0.118.0",
    "fastapi-mail>=1.5.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=7.1.0",
    "psycopg2-binary>=2.9.10",
    "pybreaker>=1.4.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "sqlparse>=0.5.3",
    "strawberry-graphql[fastapi]>=0.282.0",
    "uvicorn[standard]>=0.37.0",
    "websockets>=15.0.1",
]
